型定義のディレクトリ構造を変えた
