+++ lib/dev/api.interface.ts
@@ -1,0 +1,16 @@
+export interface zipCloudJson {
+  message: null,
+  results: [
+    {
+      address1: string,
+      address2: string,
+      address3: string,
+      kana1: string,
+      kana2: string,
+      kana3: string,
+      prefcode: string,
+      zipcode: string
+    }
+  ],
+  status: number
+}
