+++ lib/dev/api.ts
@@ -1,21 +1,5 @@
 import Axios, {AxiosResponse} from 'axios'
-
-export interface zipCloudJson {
-  message: null,
-  results: [
-    {
-      address1: string,
-      address2: string,
-      address3: string,
-      kana1: string,
-      kana2: string,
-      kana3: string,
-      prefcode: string,
-      zipcode: string
-    }
-  ],
-  status: number
-}
+import { zipCloudJson } from './api.interface'
 
 export default class Api {
   public getAddressData(zipcode: number) {
